Add new images, complete styling for sidebar and top content
The new images provide a better atmosphere for the form compared to
the previous background image. The background colors are also updated
to correspond to the image and vibe of the form page. The top portion
of the main content styling contrasts from the form itself, which is
coming together.